# refactor filtering for faster rendering