Here are 2 timelines with the before (bigger file) and after for the search "modelico". Summarizing, the before shows ~4 s of scripting and ~16 s of rendering, while the after shows ~0.3 s of scripting and ~0.4 s  rendering.

[timelines.zip](https://github.com/webjars/webjars/files/618968/timelines.zip)